requirements.txt
flask
requests
pillow 
numpy